WEB_HM_11/webhm11/pyproject.toml
[tool.poetry]
name = "webhm11"
version = "0.1.0"
description = ""
authors = ["igorgroni <igorgroni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-mail = "^1.4.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.36.0"
pytest = "^7.4.3"
sphinx = "^7.2.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
